Also allow svd patch based modifications; add some
Create a base yaml patch file for each of the six devices, including a
base.yaml that makes changes to all of them.

Remove one of the changes from svd.patch and put it into base.yaml.
Also add a bunch of write constraints (both enumerated-values and range
constraints) to the SERCOM0 SPI fields, to allow safe access.

Create a local copy of svdtools and add support to it to match on a
cluster tag (otherwise the SPI cluster of the SERCOM0 peripheral can't
be matched; see https://github.com/stm32-rs/svdtools/issues/41 for the
upstream report).